+++ src/main/java/custom/lib/HapData.java
@@ -7,10 +7,10 @@ public class HapData {
     public int start;
     public int end;
     public boolean[] states;
-    public short readCount;
+    public int readCount;
     public Strand strand;
 
-    public HapData(String _chr, int _start, int _end, boolean[] _states, short _readCount, Strand _strand) {
+    public HapData(String _chr, int _start, int _end, boolean[] _states, int _readCount, Strand _strand) {
         this.chr = _chr;
         this.start = _start;
         this.end = _end;
